chore(richtext-lexical): remove vendored @lexical/markdown fork and upgrade to lexical 0.45
Payload vendored a full copy of @lexical/markdown because, prior to lexical
0.42, importing the published package transitively pulled in @lexical/code and
its prismjs dependency, which broke type generation. Since lexical 0.42 the
markdown package depends on @lexical/code-core (no prism), so the vendored copy
is no longer needed.

- Upgrade lexical and all @lexical/* packages from 0.41.0 to 0.45.0
- Add @lexical/markdown@0.45.0 and delete the vendored copy under src/packages
- Replace the fork with a thin wrapper (lexical-proxy/@lexical-markdown.ts) that
  preserves Payload's behaviors on top of upstream: a curated transformer set
  (excluding CODE and LINK, which Payload ships as its own features) and
  shouldMergeAdjacentLines defaulting to true
- Keep Payload's normalizeMarkdown (nested code fences, table-row whitespace) as
  a dedicated module, since upstream's normalizer is not exported
- Drop the LexicalContextMenuPlugin re-export, removed upstream in lexical 0.42

Note: a markdown table cell containing only whitespace now imports as an empty
cell instead of preserving the padding spaces as a text node.
