+++ packages/richtext-lexical/src/features/blocks/server/markdown/markdownTransformer.ts
@@ -1,16 +1,16 @@
 import type { ElementNode, SerializedLexicalNode } from 'lexical'
 import type { Block } from 'payload'
 
+import {
+  type MultilineElementTransformer,
+  type TextMatchTransformer,
+  type Transformer,
+} from '@lexical/markdown'
 import { $parseSerializedNode } from 'lexical'
 
 import type { NodeWithHooks } from '../../../typesServer.js'
 
 import { getEnabledNodesFromServerNodes } from '../../../../lexical/nodes/index.js'
-import {
-  type MultilineElementTransformer,
-  type TextMatchTransformer,
-  type Transformer,
-} from '../../../../packages/@lexical/markdown/index.js'
 import { extractPropsFromJSXPropsString } from '../../../../utilities/jsx/extractPropsFromJSXPropsString.js'
 import { propsToJSXString } from '../../../../utilities/jsx/jsx.js'
 import { getLexicalToMarkdown } from '../../client/markdown/getLexicalToMarkdown.js'
