+++ packages/richtext-lexical/src/features/format/strikethrough/markdownTransformers.ts
@@ -1,4 +1,4 @@
-import type { TextFormatTransformer } from '../../../packages/@lexical/markdown/MarkdownTransformers.js'
+import type { TextFormatTransformer } from '@lexical/markdown'
 
 export const STRIKETHROUGH: TextFormatTransformer = {
   type: 'text-format',
