+++ packages/richtext-lexical/src/features/horizontalRule/server/markdownTransformer.ts
@@ -1,4 +1,4 @@
-import type { ElementTransformer } from '../../../packages/@lexical/markdown/MarkdownTransformers.js'
+import type { ElementTransformer } from '@lexical/markdown'
 
 import {
   $createHorizontalRuleServerNode,
