+++ packages/richtext-lexical/src/features/link/markdownTransformer.spec.ts
@@ -2,7 +2,7 @@ import { createHeadlessEditor } from '@lexical/headless'
 import { $createParagraphNode, $createTextNode, $getRoot } from 'lexical'
 import { describe, expect, it } from 'vitest'
 
-import { $convertToMarkdownString } from '../../packages/@lexical/markdown/index.js'
+import { $convertToMarkdownString } from '@lexical/markdown'
 import { AutoLinkNode } from './nodes/AutoLinkNode.js'
 import { $createLinkNode, LinkNode } from './nodes/LinkNode.js'
 import type { SerializedLinkNode } from './server/schema.js'
