+++ packages/richtext-lexical/src/features/typesServer.ts
@@ -1,3 +1,4 @@
+import type { Transformer } from '@lexical/markdown'
 import type { GenericLanguages, I18nClient } from '@payloadcms/translations'
 import type { JSONSchema4 } from 'json-schema'
 import type {
@@ -27,7 +28,6 @@ import type {
 } from 'payload'
 
 import type { ServerEditorConfig } from '../lexical/config/types.js'
-import type { Transformer } from '../packages/@lexical/markdown/index.js'
 import type { LexicalRichTextField } from '../types/index.js'
 import type { ElementNodeSchemaFn } from '../types/jsonSchemaHelpers.js'
 import type { BaseClientFeatureProps } from './typesClient.js'
