--- packages/richtext-lexical/src/lexical-proxy/@lexical-react/LexicalContextMenuPlugin.tsx
@@ -1,1 +1,0 @@
-export * from '@lexical/react/LexicalContextMenuPlugin'
