+++ packages/richtext-lexical/src/lexical/plugins/MarkdownShortcut/index.tsx
@@ -1,9 +1,9 @@
 'use client'
 
+import { registerMarkdownShortcuts } from '@lexical/markdown'
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
 import * as React from 'react'
 
-import { registerMarkdownShortcuts } from '../../../packages/@lexical/markdown/MarkdownShortcuts.js'
 import { useEditorConfigContext } from '../../config/client/EditorConfigProvider.js'
 
 export const MarkdownShortcutPlugin: React.FC = () => {
