+++ packages/richtext-lexical/src/lexical/utils/markdown/createBlockNode.ts
@@ -1,6 +1,5 @@
+import type { ElementTransformer } from '@lexical/markdown'
 import type { ElementNode } from 'lexical'
-
-import type { ElementTransformer } from '../../../packages/@lexical/markdown/MarkdownTransformers.js'
 
 export const createBlockNode = (
   createNode: (match: Array<string>) => ElementNode,
