+++ test/lexical-mdx/tableJson.ts
@@ -162,19 +162,13 @@ export const tableJson = {
           rowSpan: 1,
         },
         {
+          // A table cell that contains only whitespace imports as an empty
+          // paragraph. The upstream @lexical/markdown normalizer (used since the
+          // vendored fork was removed) collapses whitespace-only lines, so the
+          // padding spaces no longer survive as a text node.
           children: [
             {
-              children: [
-                {
-                  detail: 0,
-                  format: 0,
-                  mode: 'normal',
-                  style: '',
-                  text: '                         ',
-                  type: 'text',
-                  version: 1,
-                },
-              ],
+              children: [],
               direction: null,
               format: '',
               indent: 0,
